Data transformers
These are used to re-structure the original data and maintain / sum the cost values by a series of columns and structures.

There are simple transforms, such as ProjectByPeriod, which flattens the data to single level array, adds in the ${project} meta value and accumulates the costs for each month

There are then multi level versions, like ProjectAndDimensionByPeriod, which groups the costing data not just by date, but also by ‘Dimension’ (AWS Service).

All of the transforms return data in a flat array, adding values as required (so ProjectByPeriod adds the Project name etc)